+++ src/MessageBase.cpp
@@ -31,7 +31,7 @@ bool MessageBase::onRead(const TypeComponentUnit& source, const uint8_t *buffer,
 //    LOGS_E(getHost(), "%ld : Data received, count : %3d, bufPtr : %s", iter++,
 //           nRead, Util::hex2str(buffer, nRead).c_str());
 
-    uint32_t minContDataLength;
+    size_t minContDataLength;
     size_t remaining = 0;
 
     auto *bufPtr = buffer;
@@ -286,7 +286,7 @@ bool MessageBase::onWrite(const TypeComponentUnit& target, MSG_HEADER blockType,
         tmpBufPos = 0;
     }
 
-    block.set(blockType, size);
+    block.set(blockType, (uint32_t)size);
 
     block.fillBuffer(tmpBuf + tmpBufPos);
 
@@ -399,7 +399,7 @@ bool MessageBase::writeBinary(const TypeComponentUnit& target,
 
     do {
 
-        int count = std::fread(tmpBuf, 1, std::min((size_t)TMP_BUFFER_SIZE, remaining), file);
+        size_t count = std::fread(tmpBuf, 1, std::min((size_t)TMP_BUFFER_SIZE, remaining), file);
 
         onWrite(target, MSGHEADER_MAX, tmpBuf, count);
